+++ packages/full_screen_dialog_manager/lib/src/dismissible/full_screen_dismissible.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_easy_dialogs/flutter_easy_dialogs.dart';
 import 'package:full_screen_dialog_manager/src/manager/full_screen_dialog_manager.dart';
 
-part 'gesture.dart';
+part 'tap.dart';
 part 'none.dart';
 
 /// This is an implementation of [EasyDialogDismissible] that is specific to
@@ -14,10 +14,10 @@ abstract class FullScreenDismissible extends EasyDialogDismissible {
   const FullScreenDismissible({super.onDismissed});
 
   /// Simple gesture tap dismiss.
-  const factory FullScreenDismissible.gesture({
+  const factory FullScreenDismissible.tap({
     HitTestBehavior behavior,
     OnEasyDismissed? onDismissed,
-  }) = _Gesture;
+  }) = _Tap;
 
   const factory FullScreenDismissible.none() = _None;
 }