+++ packages/positioned_dialog_manager/lib/src/dismissible/positioned_dismissible.dart
@@ -4,9 +4,9 @@ import 'package:flutter_easy_dialogs/flutter_easy_dialogs.dart';
 import 'package:positioned_dialog_manager/src/easy_dialog_position.dart';
 import 'package:positioned_dialog_manager/src/manager/positioned_dialog_manager.dart';
 
-part 'tap.dart';
+part 'animated_tap.dart';
 part 'swipe.dart';
-part 'gesture.dart';
+part 'tap.dart';
 part 'none.dart';
 
 /// Dismissible that is used by [PositionedDialogManager].
@@ -15,14 +15,16 @@ abstract class PositionedDismissible extends EasyDialogDismissible<
   const PositionedDismissible({required super.onDismissed});
 
   /// Simple tap gesture dismissible.
-  const factory PositionedDismissible.gesture({
-    OnEasyDismissed? onDismissed,
-  }) = _Gesture;
-
-  /// Tap gesture but with extra `scale in` on tap down animation.
   const factory PositionedDismissible.tap({
+    HitTestBehavior? behavior,
     OnEasyDismissed? onDismissed,
   }) = _Tap;
+
+  /// Tap gesture but with extra `scale in` on tap down animation.
+  const factory PositionedDismissible.animatedTap({
+    Duration duration,
+    OnEasyDismissed? onDismissed,
+  }) = _AnimatedTap;
 
   /// Horizontal swipe dismissible.
   ///