+++ packages/positioned_dialog_manager/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.4
+
+* Renamed `PositionedDismissible.tap` to `PositionedDismissible.animatedTap`, added `duration` parameter to specify animation duration
+* Renamed `PositionedDismissible.gesture` to `PositionedDismissible.tap`, added `behavior` to specify `HitTestBehavior`
+* Renamed `DialogsMap` to `AnimationControllerBox`
+
 ## 1.0.3+2
 
 * Updated README.md